ADD: DTOs and Mappers

+++ src/main/java/com/inversionesaraujo/api/repository/CartProductRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.CartProduct;
 
-public interface CartProductRepository extends JpaRepository<CartProduct, Integer> {}
+public interface CartProductRepository extends JpaRepository<CartProduct, Long> {}
