+++ src/main/java/com/inversionesaraujo/api/repository/CartRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Cart;
 
-public interface CartRepository extends JpaRepository<Cart, Integer> {}
+public interface CartRepository extends JpaRepository<Cart, Long> {}
