+++ src/main/java/com/inversionesaraujo/api/repository/ExpenseRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Expense;
 
-public interface ExpenseRepository extends JpaRepository<Expense, Integer> {}
+public interface ExpenseRepository extends JpaRepository<Expense, Long> {}
