+++ src/main/java/com/inversionesaraujo/api/repository/OfferProductRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.OfferProduct;
 
-public interface OfferProductRepository extends JpaRepository<OfferProduct, Integer> {}
+public interface OfferProductRepository extends JpaRepository<OfferProduct, Long> {}
