+++ src/main/java/com/inversionesaraujo/api/repository/OfferRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Offer;
 
-public interface OfferRepository extends JpaRepository<Offer, Integer> {}
+public interface OfferRepository extends JpaRepository<Offer, Long> {}
