+++ src/main/java/com/inversionesaraujo/api/repository/OrderVarietyRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.OrderVariety;
 
-public interface OrderVarietyRepository extends JpaRepository<OrderVariety, Integer> {}
+public interface OrderVarietyRepository extends JpaRepository<OrderVariety, Long> {}
