+++ src/main/java/com/inversionesaraujo/api/repository/ProductImageRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.ProductImage;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {}
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {}
