+++ src/main/java/com/inversionesaraujo/api/repository/TuberRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.inversionesaraujo.api.model.Tuber;
 
-public interface TuberRepository extends JpaRepository<Tuber, Integer> {}
+public interface TuberRepository extends JpaRepository<Tuber, Long> {}
